Working towards parsing of a Let expression.
Realising that our DFA is too simple; need certain lexemes to appear in different places in the machine.
E.g. our Let state, we should ignore assignment as it is inferred in the current model (if children is not empty, there is assignment).
This means we need a special "let-assign" state in the machine that won't push the assignment operator on to the node stack.
We wouldn't want the assign operator to be ignored in other contexts (e.g. reassignment, "foo = bar")
Suspect that this will give our DFA the power to define much stricter language rules.
I.e. we can define different parse behaviour based on previous transitions around the machine.

The above assumes our Let parse tree is:

Let
 |
 -- Identifier
 |
 -- Type identifier
 |
 -- []Node // after assignment

Rather than (the potentially simpler?)

Operator: =
 |
 -- Let
 |   |
 |   -- Identifier
 |   |
 |   -- Type identifier
 |
 -- []Node // after assignment

The latter would imply we could use Let statements elsewhere, which isn't desirable?